helpers: improve dwim behavior
Treat the item in single-item arrays as top-level nouns, to have parity
with [noun] cell notation meaning just noun.

Produce the zero atom for null, in line with common usage.

Throw an error if the "meaning"/intended noun is unclear, instead of
silently producing undefined. This should make dwim issues easier to
catch.