noun, helpers: remove Atom.valueOf(), add helpers
Atom.valueOf() would give you the atom as either a number or a string.
This kind of "un-steady" affordance felt out of place on the Atom class
proper. We remove it here, and replace internal calls to it with
Number() construction or Atom equality checks as appropriate.

To replace it, we add two new conversion helpers. One for converting
Atoms that fit within 32 bits (that crashes if the atom is too large),
and one for converting Atoms into their string representations. The
pre-existing numb() implementation now matches the behavior of those as
well.