bits: improve cue perf by orders of magnitude
Cueing big or many-celled nouns turned out to be very slow. Here we
tackle the worst offenders:

bitLength() was "walking" the full bigint bit-for-bit. Turns out that
just stringifying it into binary and getting the length of that string
is faster than a manual loop.

bigIntFromStringWithRadix() was manually constructing the bigint value
step-by-step instead of using the built-in constructor. We now use the
built-in constructor for well-known radices.

cut() uses a bunch of logic so it can remain agnostic to the bloq size
it's working with. But the bloq size 0 is very common during cue and can
be trivially implemented with native bitshift operations.

(We could probably get further speedups by applying that last approach
to chop() and maybe other internals also.)

The 5kb jam test case used to take ~7 minutes. Now it takes less than .2
seconds.