bigint: further optimize bitlength calculation
This was still by far the slowest part of big cues, because we need to
measure big atoms often. Here we take an overengineered approach.

We add a test with a ~160kb jam buffer. It used to take nearly two
minutes, now takes about five seconds. Still not super fast, but
we'll get there.