serial: make cue operate on byte buffer & numbers
Cue's core operation is "read bits from jam buffer". We had naively
implemented this to operate on the Atom representing the jam buffer,
and use Atoms for all intermediate steps in these operations.
Turns out that this is quite slow.

Turning the jam buffer into a real byte buffer lets us look at it
through the DataView interface, dramatically increasing the performance
of reads.

Additionally, numbers are generally faster than bigints (which back
Atoms), so we reimplement, so we use them wherever we can, which is
almost everywhere except the atoms we read out of the jam.

We also now expose an interface for cueing byte buffers directly, since
that's often the initial shape of data received over the network.