noun: properly type tree addressing functions
The fragmenter() should produce a function that takes and produces a
Noun. The at() functions should produce a Noun, because they call a
function produced by the fragmenter().

(Having .at() on Atoms is a little bit weird, but lets you call it on
opaque Nouns. It'll crash in the same way that it would crash on shallow
Cells or other "miss" cases.)