helpers: experimental "face addressing" utility
Introduces experimental utilities for generating axes from a faces array
that structurally represents the noun it's addressing, and then grabbing
parts of that noun using those faces.

Does not support nested faces. Here as an experimental helper, need to
see this used in the wild to properly judge its utility/api design.
Ultimately, it would be good if we could `myNoun.applyFaces` or
whatever, and afterwards `myNoun.at('face')` or even
`myNoun.at('face.etc')`. For the latter, and for nested faces in
general, we would probably want to support masks with nested faces, and
propagate those into the inner nouns. Could use the mask for enhanced
pretty-printing, too...