noun: use modern mug algorithm
The mug implementation here was two iterations behind. Here, we replace
it with the modern murmur3-based implementation.

We crib @garycourt's murmur3 implementation. It was also used in
urbit-ob and remains in use for aura-js's planet name scrambling
implementation. We make some small modifications to it so we can pass it
native bigints and an optional bytelength argument in case of leading
zeroes.

We add tests for the new mug-related stdlib primitives and update all
mug hash checks in existing tests.